requirements.txt
# MAZLUM - Mac-first, minimal dependencies
# Core LLM providers
anthropic>=0.40.0
google-generativeai>=0.8.0

# Web server
fastapi>=0.115.0
uvicorn>=0.32.0
websockets>=13.0

# Telegram (Bot API - production-primary)
python-telegram-bot>=21.0

# Database (read-only intelligence)
sqlalchemy>=2.0.0
# Install the driver you need:
# pip install psycopg2-binary   # PostgreSQL
# pip install pymysql           # MySQL/MariaDB
# pip install pyodbc            # MSSQL

# Search
duckduckgo-search>=7.0.0

# Telegram User API (real-time monitoring)
telethon>=1.35.0

# Voice (optional, install if needed)
# pyaudio>=0.2.14

# Document creation
python-docx>=1.1.0
openpyxl>=3.1.0

# Image / Vision
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
